Refactor and break up NerdDice module
Fix RuboCop violations introduced in the previous commit by breaking up
the NerdDice module source code file into a hierarchy of smaller files
that are required by the base file. Refactor the roll_ability_scores
method interpret_ability_score_options subroutine to resolve branch
condition violation.

Also move spec files for NerdDice class methods to class_methods
subfolder in spec/nerd_dice and increase benchmark for securerandom.

Completes NerdDice backlog item:
Write roll_ability_scores method
