+++ spec/nerd_dice/class_methods/execute_die_roll_spec.rb
