+++ spec/nerd_dice/class_methods/refresh_seed_spec.rb
