+++ spec/nerd_dice/class_methods/roll_ability_scores_spec.rb
