+++ spec/nerd_dice/class_methods/roll_dice_spec.rb
