+++ spec/nerd_dice/class_methods/total_dice_spec.rb
